+++ ThirdPartyAdapters/dtexchange/dtexchange/src/test/kotlin/com/google/ads/mediation/fyber/FyberMediationAdapterTest.kt
@@ -36,6 +36,7 @@ import com.google.ads.mediation.adaptertestkit.createMediationRewardedAdConfigur
 import com.google.android.gms.ads.AdError
 import com.google.android.gms.ads.AdFormat
 import com.google.android.gms.ads.MobileAds
+import com.google.android.gms.ads.RequestConfiguration
 import com.google.android.gms.ads.mediation.InitializationCompleteCallback
 import com.google.android.gms.ads.mediation.MediationAdLoadCallback
 import com.google.android.gms.ads.mediation.MediationBannerAd
@@ -98,6 +99,14 @@ class FyberMediationAdapterTest {
   // region Setup
   @Before
   fun setUp() {
+    val requestConfiguration =
+      RequestConfiguration.Builder()
+        .setTagForChildDirectedTreatment(
+          RequestConfiguration.TAG_FOR_CHILD_DIRECTED_TREATMENT_UNSPECIFIED
+        )
+        .setTagForUnderAgeOfConsent(RequestConfiguration.TAG_FOR_UNDER_AGE_OF_CONSENT_UNSPECIFIED)
+        .build()
+    MobileAds.setRequestConfiguration(requestConfiguration)
     adapter = FyberMediationAdapter()
     mockInneractiveAdManager = mockStatic(InneractiveAdManager::class.java)
     FyberSdkWrapper.delegate = mockSdkWrapper
@@ -161,6 +170,11 @@ class FyberMediationAdapterTest {
   // region Initialization Tests
   @Test
   fun initialize_initializesInneractiveAdManager() {
+    val requestConfiguration =
+      RequestConfiguration.Builder()
+        .setTagForChildDirectedTreatment(RequestConfiguration.TAG_FOR_CHILD_DIRECTED_TREATMENT_TRUE)
+        .build()
+    MobileAds.setRequestConfiguration(requestConfiguration)
     val initializationParameters = createMediationConfiguration(AdFormat.BANNER, serverParameters)
     adapter.initialize(
       activity,
@@ -169,6 +183,7 @@ class FyberMediationAdapterTest {
     )
 
     mockInneractiveAdManager.verify {
+      InneractiveAdManager.currentAudienceAppliesToCoppa()
       InneractiveAdManager.initialize(eq(activity), eq(AdapterTestKitConstants.TEST_APP_ID), any())
     }
   }
@@ -183,6 +198,10 @@ class FyberMediationAdapterTest {
       listOf(initializationParameters),
     )
 
+    mockInneractiveAdManager.verify(
+      { InneractiveAdManager.currentAudienceAppliesToCoppa() },
+      never(),
+    )
     mockInneractiveAdManager.verify {
       InneractiveAdManager.initialize(
         eq(activity),
@@ -250,6 +269,11 @@ class FyberMediationAdapterTest {
 
   @Test
   fun initialize_withMultipleAppIds_initializesOnlyOnceAndInvokesOnInitializationSucceeded() {
+    val requestConfiguration =
+      RequestConfiguration.Builder()
+        .setTagForUnderAgeOfConsent(RequestConfiguration.TAG_FOR_UNDER_AGE_OF_CONSENT_TRUE)
+        .build()
+    MobileAds.setRequestConfiguration(requestConfiguration)
     val serverParameters1 = bundleOf(FyberMediationAdapter.KEY_APP_ID to TEST_APP_ID_1)
     val serverParameters2 = bundleOf(FyberMediationAdapter.KEY_APP_ID to TEST_APP_ID_2)
     val initializationParameters1 = createMediationConfiguration(AdFormat.BANNER, serverParameters1)
@@ -262,12 +286,13 @@ class FyberMediationAdapterTest {
       listOf(initializationParameters1, initializationParameters2),
     )
 
+    mockInneractiveAdManager.verify() { InneractiveAdManager.currentAudienceAppliesToCoppa() }
     mockInneractiveAdManager.verify(
       { InneractiveAdManager.initialize(eq(activity), eq(TEST_APP_ID_2), listener.capture()) },
       times(1),
     )
     mockInneractiveAdManager.verify(
-      { InneractiveAdManager.initialize(eq(activity), eq(TEST_APP_ID_1), listener.capture()) },
+      { InneractiveAdManager.initialize(eq(activity), eq(TEST_APP_ID_1), any()) },
       never(),
     )